requirements.txt
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
Bottleneck @ file:///opt/conda/conda-bld/bottleneck_1657175564434/work
category-encoders==1.3.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.2
contourpy==1.0.7
cornac==1.14.2
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
exceptiongroup==1.1.0
executing==1.2.0
filelock==3.9.0
fonttools==4.39.0
huggingface-hub==0.13.0
hypothesis==6.68.2
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
ipykernel==6.21.3
ipython==8.11.0
jedi==0.18.2
Jinja2==3.0.3
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
kiwisolver==1.4.4
lightfm==1.16
lightgbm==3.3.5
llvmlite==0.39.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
memory-profiler==0.61.0
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186066731/work
mkl-service==2.4.0
mpmath==1.3.0
mypy-extensions==1.0.0
nest-asyncio==1.5.6
nltk==3.8.1
numba==0.56.4
numexpr @ file:///croot/numexpr_1668713893690/work
numpy @ file:///croot/numpy_and_numpy_base_1672336185480/work
packaging==23.0
pandas==1.5.3
pandera==0.13.4
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.1.0
powerlaw==1.5
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.6
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz @ file:///croot/pytz_1671697431263/work
PyYAML==5.4.1
pyzmq==25.0.0
recommenders==1.1.1
regex==2022.10.31
requests==2.28.2
retrying==1.3.4
scikit-learn==1.0.2
scikit-surprise==1.1.3
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
sortedcontainers==2.4.0
stack-data==0.6.2
statsmodels==0.13.5
threadpoolctl==3.1.0
tokenizers==0.13.2
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.26.1
typing-inspect==0.8.0
typing_extensions==4.5.0
urllib3==1.26.14
wcwidth==0.2.6
wrapt==1.15.0
zipp==3.15.0
